requirements.txt
selenium-recaptcha-solver
beautifulsoup4
# selenium==4.8.0